Fix spec namespaces
There were some not needed `require_relative` calls if the `spec` folder
were ordered according to the `lib` folder.

So, I removed them and organized the test suite accordingly the classes at library
